use raylet signal handling in plasma store when running plasma store as a thread
