+++ src/ray/raylet/raylet.cc
@@ -90,6 +90,7 @@ void Raylet::Start() {
 }
 
 void Raylet::Stop() {
+  object_manager_.Stop();
   RAY_CHECK_OK(gcs_client_->Nodes().UnregisterSelf());
   acceptor_.close();
 }
